+++ changes.md
@@ -4,4 +4,4 @@
 * compability with themes and (block)editor
 * sgpx (from WP GPX Maps) does not work with some themes (like TT2)
 * new shortcode: choropleth - https://github.com/timwis/leaflet-choropleth
-* new shortcodes leaflet-optiongroups and leaflet-featuregroups (Leaflet.FeatureGroup.SubGroup) (todo: documentation)
+* new shortcodes leaflet-optiongroup and leaflet-featuregroup (Leaflet.FeatureGroup.SubGroup)